instaclone-api/requirements.txt
anyio==3.6.1
bcrypt==3.2.2
beanie==1.11.6
cffi==1.15.1
click==8.1.3
cryptography==37.0.4
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.78.0
fastapi-users==10.1.1
fastapi-users-db-beanie==1.1.1
h11==0.13.0
idna==3.3
makefun==1.13.1
motor==3.0.0
multidict==6.0.2
passlib==1.7.4
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
pymongo==4.1.1
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
typing_extensions==4.3.0
uvicorn==0.18.2
yarl==1.7.2
anyio==3.6.1
bcrypt==3.2.2
beanie==1.11.6
cffi==1.15.1
click==8.1.3
cryptography==37.0.4
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.78.0
fastapi-users==10.1.1
fastapi-users-db-beanie==1.1.1
h11==0.13.0
idna==3.3
makefun==1.13.1
motor==3.0.0
multidict==6.0.2
passlib==1.7.4
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
pymongo==4.1.1
python-dotenv==0.20.0
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
typing_extensions==4.3.0
uvicorn==0.18.2
yarl==1.7.2
